change ./tools to ./util, remove Openwrist/MEII references in README, fix all compiler warnings in examples

+++ examples/ex_udp.cpp
@@ -69,7 +69,7 @@ void client(int iterations, int bytes, const IpAddress& remote_address) {
         clock.restart();
         socket.send(packet, remote_address, 55002);
         socket.receive(packet, sender, port);
-        pings.push_back(clock.get_elapsed_time().as_microseconds());
+        pings.push_back((double)clock.get_elapsed_time().as_microseconds());
     }
     packet.clear();
     msg = "done";
@@ -77,10 +77,10 @@ void client(int iterations, int bytes, const IpAddress& remote_address) {
     socket.send(packet, remote_address, 55002);
     print("Done");
     // Print results
-    int mean_ping = mean(pings);
-    int std_ping = stddev_p(pings);
-    int min_ping = min(pings);
-    int max_ping = max(pings);
+    int mean_ping = (int)mean(pings);
+    int std_ping = (int)stddev_p(pings);
+    int min_ping = (int)min(pings);
+    int max_ping = (int)max(pings);
     print("Avg Ping: " + stringify(mean_ping) + " +/- " + stringify(std_ping) + " usec");
     print("Min Ping: " + stringify(min_ping) + " usec");
     print("Max Ping: " + stringify(max_ping) + " usec");
